Move pipeline::Config to separate mod
