+++ src/pipeline/mod.rs
@@ -1,9 +1,11 @@
+pub mod config;
 pub mod deferred;
 pub mod glow;
 pub mod render_pass;
 pub mod shaders;
 pub mod shadow;
 
+pub use config::Config;
 pub use render_pass::{CompositionPassComponent, RenderPassComponent, ScenePassComponent};
 
 use coarse_prof::profile;
@@ -21,29 +23,6 @@ use crate::{
 use deferred::DeferredShading;
 use glow::Glow;
 use shadow::ShadowMapping;
-
-#[derive(Debug, Clone)]
-pub struct Config {
-    pub shadow_mapping: Option<shadow::Config>,
-    pub deferred_shading: Option<deferred::Config>,
-    pub glow: Option<glow::Config>,
-    pub hdr: Option<f32>,
-    pub gamma_correction: Option<f32>,
-    pub fxaa: Option<fxaa::Config>,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            shadow_mapping: Some(Default::default()),
-            deferred_shading: Some(Default::default()),
-            glow: Some(Default::default()),
-            hdr: None,
-            gamma_correction: Some(2.2),
-            fxaa: Some(Default::default()),
-        }
-    }
-}
 
 struct Components {
     shadow_mapping: Option<ShadowMapping>,
